requirements.txt
beautifulsoup4>=4.12.0
lxml>=5.0.0
openai>=1.6.0
flask>=3.0.0
pillow>=10.0.0
werkzeug>=3.0.0
